+++ crypt-data/src/test/java/de/alpharogroup/crypto/factories/KeyPairFactoryTest.java
@@ -1,13 +1,16 @@
 package de.alpharogroup.crypto.factories;
 
-import java.io.File;
-import java.security.PrivateKey;
-import java.security.Security;
+import static org.testng.Assert.assertNotNull;
 
-import org.bouncycastle.jce.provider.BouncyCastleProvider;
+import java.io.File;
+import java.security.KeyPair;
+import java.security.PrivateKey;
+import java.security.PublicKey;
+
 import org.testng.annotations.Test;
 
 import de.alpharogroup.crypto.key.reader.PrivateKeyReader;
+import de.alpharogroup.crypto.key.reader.PublicKeyReader;
 import de.alpharogroup.file.search.PathFinder;
 
 /**
@@ -23,13 +26,17 @@ public class KeyPairFactoryTest {
 	 */
 	@Test
 	public void testProtectPrivateKeyWithPassword() throws Exception {
-		final File publickeyDerDir = new File(PathFinder.getSrcTestResourcesDir(), "der");
-		final File privatekeyDerFile = new File(publickeyDerDir, "private.der");
 
-		Security.addProvider(new BouncyCastleProvider());
+		final File publickeyDerDir = new File(PathFinder.getSrcTestResourcesDir(), "der");
+		final File publickeyDerFile = new File(publickeyDerDir, "public.der");
+		final File privatekeyDerFile = new File(publickeyDerDir, "private.der");
 
 		final PrivateKey privateKey = PrivateKeyReader.readPrivateKey(privatekeyDerFile);
 
+		final PublicKey publicKey = PublicKeyReader.readPublicKey(publickeyDerFile);
+
+		final KeyPair keyPair = KeyPairFactory.newKeyPair(publicKey, privateKey);
+		assertNotNull(keyPair);
 	}
 
 }