+++ CHANGELOG.md
@@ -6,7 +6,8 @@ Version 4.25.0-SNAPSHOT
 
 ADDED:
  
-- 
+- protect and encrypt private key with password
+- write protected private key with password to a file
 
 CHANGED: